requirements.txt
aiogram>=3.0.0,<4.0.0
SQLAlchemy>=2.0,<3.0
aiosqlite>=0.19,<1.0
python-dotenv>=1.0,<2.0
alembic>=1.12,<2.0
